* semantics/asfc/src/lex-cons.c: explicit cast to (int) in PT_makeTreeChar

+++ asf/semantics/asfc/src/lex-cons.c
@@ -10,7 +10,7 @@ static PT_Tree constructorCharToLexicalChar(ASF_CHAR ch)
 {
   const char *str = ASF_getCHARString(ch);
 
-  return PT_makeTreeChar(str[1]);
+  return PT_makeTreeChar((int) str[1]);
 }
 
 static PT_Tree constructorVarToLexicalVar(ASF_CHAR var)
